fix dropping items to editor3
- from related items widget
- from metasearch

SDANSA-209 SDANSA-210
