+++ client/ansa/widgets/ansa-related-widget.html
@@ -12,22 +12,22 @@
     </div>
     <div class="widget" sd-loading="!related.items">
         <ul class="simple-list simple-list--comfortable simple-list--dotted simple-list--no-padding" sd-loading="!related.items">
-            <li ng-repeat="relatedItem in related.items track by $index"
-                class="simple-list__item" sd-draggable-item ng-init="isCollapsed = relatedItem.type === 'text'">
+            <li ng-repeat="item in related.items track by $index"
+                class="simple-list__item" sd-draggable-item ng-init="isCollapsed = item.type === 'text'">
                 <div class="basic item-block">
-                    <i sd-filetype-icon data-item="relatedItem"></i>
-                    <span class="keyword" ng-click="isCollapsed = !isCollapsed">{{ relatedItem.headline }}</span>
-                    <div class="headline"><time sd-datetime data-date="relatedItem.versioncreated"></time></div>
+                    <i sd-filetype-icon data-item="item"></i>
+                    <span class="keyword" ng-click="isCollapsed = !isCollapsed">{{ item.headline }}</span>
+                    <div class="headline"><time sd-datetime data-date="item.versioncreated"></time></div>
                     <div class="collapsed-content">
-                        <div ng-if="relatedItem.body_html && relatedItem.type !== 'picture'" ng-hide="isCollapsed">
-                            <div sd-html-preview="relatedItem.body_html"></div>
+                        <div ng-if="item.body_html && item.type !== 'picture'" ng-hide="isCollapsed">
+                            <div sd-html-preview="item.body_html"></div>
                         </div>
 
-                        <div ng-if="relatedItem.type === 'picture'">
-                            <div sd-item-rendition data-item="relatedItem" data-rendition="thumbnail"></div>
+                        <div ng-if="item.type === 'picture'">
+                            <div sd-item-rendition data-item="item" data-rendition="thumbnail"></div>
                             <div ng-if="related.allowAddMedia()" class="set-related-btn-group">
-                                <button class="btn btn--small" ng-click="related.setFeatured(relatedItem)" translate>Set as Featured</button>
-                                <button class="btn btn--small" ng-click="related.addToGallery(relatedItem)" translate>Add to Gallery</button>
+                                <button class="btn btn--small" ng-click="related.setFeatured(item)" translate>Set as Featured</button>
+                                <button class="btn btn--small" ng-click="related.addToGallery(item)" translate>Add to Gallery</button>
                             </div>
                         </div>
                     </div>
